patch to remove msvc warnings
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@2496 d0c010d8-bb82-477e-a075-112140230451

+++ src/CGarminPolygon.cpp
@@ -314,7 +314,7 @@ quint32 CGarminPolygon::decode2(qint32 iCenterLon, qint32 iCenterLat, quint32 sh
     //     qDebug() << ">>" << bs_len << bytes_total << (pEnd - pStart);
 
     //     assert((pEnd - pStart) >= bytes_total);
-    if((pEnd - pStart) < bytes_total)
+    if(((quint32)(pEnd - pStart)) < bytes_total)
     {
         return (pEnd - pStart);
     }
